+++ angular_ast/lib/src/visitor.dart
@@ -8,6 +8,7 @@ export 'visitors/desugar_visitor.dart';
 export 'visitors/expression_parser_visitor.dart';
 export 'visitors/humanizing.dart';
 export 'visitors/identity.dart';
+export 'visitors/recursive.dart';
 
 /// A visitor for [TemplateAst] trees that may process each node.
 ///